Today and Upcoming Views + Local storage
The tasks can now be scheduled to when you want to do them. This puts them in either the today view - or the upcoming view which shows tasks scheduled or due in the next week. Tasks now also disappear when checked off.

Additionally, everything is now saved in the users local storage so  when the website is closed and re-opened (or refreshed) everything remains.